pyproject.toml
[build-system]
requires = ["setuptools >= 75.3"]
build-backend = "setuptools.build_meta"

[project]
name = "neuro"
readme = "README.md"
version = "1.2.1"
dependencies = [
    "bs4>=0.0.1",
    "click>=8.1.0",
    "deepdiff>=6.7.1",
    "halo>=0.0.31",
    "lxml>=5.1.0",
    "pillow>=10.2.0",
    "psutil>=5.9.7",
    "pyexiv2>=2.15.3",
    "pytest>=7.4.4",
    "python-dotenv>=1.0.1",
    "python-magic>=0.4.27",
    "pytz>=2024.1",
    "requests>=2.31.0",
    "tqdm>=4.66.1",
    "urllib3>=2.1.0"
]

[project.scripts]
neuro = "neuro.tools.terminal.cli:cli"

[project.urls]
repository = "https://github.com/neuroforest/neuro"

[tool.pytest.ini_options]
markers = [
    "integration: integration test"
]

[tool.setuptools.packages.find]
include = ["neuro*"]
exclude = ["build*", "tests*"]
